Fix more merge conflicts
